LayoutTests/platform/ios/accessibility/dialog-slotted-content-expected.txt

This test ensures that slotted content inside dialogs is accessible.

WebArea AXLabel:
    ApplicationDialog AXLabel: web dialog
        Heading AXLabel: Hello world
            StaticText AXLabel: Hello world
        Button AXLabel: Non-slotted button
        Paragraph AXLabel:
            StaticText AXLabel: Slotted paragraph text.
        WebCoreLink AXLabel: Slotted link
            StaticText AXLabel: Slotted link
        TextField AXLabel:
            Generic AXLabel:
        Button AXLabel: Slotted Button

PASS successfullyParsed is true

TEST COMPLETE
Slotted paragraph text.

Slotted link    Slotted Button
